requirements.txt
Flask==1.1.2
flask-restful==0.3.8
pycrypto==2.6.1